move rcpsp with blocking resource to its own module
