+++ src/discrete_optimization/rcpsp_blocking_resource/problem_with_blocking.py
